requirements.txt
streamlit>=1.30.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.1.0
